chore: Update provider source and version, fix typo
This commit updates the Terraform provider source from
`registry.terraform.io/abrha/abrha` to `abrhacom/abrha` in several
examples and the README.md.

It also changes the version constraint from `>= 1.0.0` to `~> 1.0` in
the examples.

Additionally, this commit fixes a typo in the vm.md documentation,
changing `monthly` to `monthday` in the `backup_policy` description.