docs: Update provider source, version, and variable name
This commit updates the provider source from "abrha/abrha" to
"abrhacom/abrha" and the version constraint from "~> 2.0" to "~> 1.0"
in the documentation. It also changes the variable name from
"do_token" to "abrha_token" to align with the provider's naming
convention. Additionally, the documentation is updated to reflect
that the required token is the Abrha API token instead of the DO API
token. This ensures that the documentation accurately reflects the
current state of the Abrha provider.